+++ packages/starlight-view-modes/libs/shared.ts
@@ -1,0 +1,1 @@
+export const currentModeKey = Symbol.for("starlight-view-modes-current-mode");