+++ packages/starlight-view-modes/middleware.ts
@@ -1,12 +1,18 @@
 import { defineRouteMiddleware } from "@astrojs/starlight/route-data";
+
+import { currentModeKey } from "./libs/shared";
 import { getCurrentMode } from "./libs/sidebar";
 
 export const onRequest = defineRouteMiddleware((context) => {
   const { starlightRoute } = context.locals;
-  const { id, sidebar, pagination } = starlightRoute;
+  const { id, sidebar, pagination, siteTitleHref } = starlightRoute;
 
   const currentMode = getCurrentMode(id, sidebar, pagination);
   starlightRoute.sidebar = currentMode.sidebar;
   starlightRoute.pagination = currentMode.pagination;
-  starlightRoute.currentMode = currentMode.mode;
+
+  starlightRoute[currentModeKey] = currentMode.mode;
+
+  if (currentMode.mode !== "default")
+    starlightRoute.siteTitleHref = `/${currentMode.mode}${siteTitleHref}`;
 });